rename some authentication files
