+++ src/back-end/models/userSchema.js
@@ -15,7 +15,7 @@ const userSchema = new Schema({
    role: {
       type: String,
       default: "user",
-      enum: ['user', 'admin']
+      enum: ['user', 'vendor']
    },
 });
