+++ src/back-end/routes/users.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-var userController = require("../controllers/userController");
+var userController = require("../controllers/authController");
 
 /* GET users listing. */
 router.get('/', function(req, res, next) {
