+++ src/back-end/tests/authTest.md
@@ -9,7 +9,7 @@ URL: http://localhost:3000/api/auth/signup
 {
   "username": "newuser@email.com",
   "password": "newpassworD123",
-  "role": "admin"
+  "role": "vendor"
 }
 ```
